requirements.txt
Django==3.0.7
Pillow==8.0.0
django-admin-sortable2==0.7.6
django-tinymce==3.0.2
environs==8.0.0